refactor diary entry controllers to make tests pass
